Refactor upload data contracts down to repository layer
- Created UploadQueryContract, UploadUpdateContract, and UploadedItemResult in the repository package.
- Refactored UploadRepository and UploadRepositoryImpl to use these layer-appropriate types instead of service types.
- Updated UploadCoordinator to map service types to repository types at the boundary, ensuring response payloads are preserved.
- Updated UploadRepositoryImplTest to use the new contracts and removed unused imports.
